backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain-groq
duckduckgo-search
